+++ src/test/java/com/moment/the/TheApplicationTests.java
@@ -3,11 +3,14 @@ package com.moment.the;
 import com.moment.the.advice.exception.UserAlreadyExistsException;
 import com.moment.the.advice.exception.UserNotFoundException;
 import com.moment.the.domain.AdminDomain;
+import com.moment.the.domain.ImprovementDomain;
 import com.moment.the.dto.AdminDto;
+import com.moment.the.dto.ImprovementDto;
 import com.moment.the.dto.SignInDto;
 import com.moment.the.repository.AdminRepository;
 import com.moment.the.service.AdminService;
 import com.moment.the.service.AdminServiceImpl;
+import com.moment.the.service.ImprovementService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -33,13 +36,6 @@ import static org.junit.jupiter.api.Assertions.*;
 @SpringBootTest
 @Transactional
 class TheApplicationTests {
-
-	// test 편의를 위한 회원가입 매서드
-	void adminSignUp(String adminId, String password, String adminName) throws Exception {
-		AdminDto adminDto = new AdminDto(adminId, password, adminName);
-		adminService.signUp(adminDto);
-	}
-
 	@AfterEach
 	public void dataClean(){
 		adminRepository.deleteAll();
@@ -264,5 +260,44 @@ class TheApplicationTests {
 
 		// When logout
 		adminServiceImpl.logout();
+	}
+
+	//    // test 편의를 위한 회원가입 매서드
+    void adminSignUp(String adminId, String password, String adminName) throws Exception {
+        AdminDto adminDto = new AdminDto(adminId, password, adminName);
+        adminService.signUp(adminDto);
+    }
+
+
+//    //test 편의를 위한 로그인 매서드
+    AdminDomain adminLogin(String adminId, String password) throws Exception {
+        AdminDomain adminDomain = adminRepository.findByAdminId(adminId);
+        UsernamePasswordAuthenticationToken token = new UsernamePasswordAuthenticationToken(
+                adminDomain.getAdminId(),
+                adminDomain.getAdminPwd(),
+                List.of(new SimpleGrantedAuthority("ROLE_USER")));
+        SecurityContext context = SecurityContextHolder.getContext();
+        context.setAuthentication(token);
+
+        return adminDomain;
+    }
+
+	@Autowired
+	private ImprovementService improvementService;
+
+	@Test
+	void 개선사례_작성() throws Exception {
+		//Given
+		ImprovementDto improvementDto = new ImprovementDto();
+		improvementDto.setImproveHeader("hello world");
+		improvementDto.setImproveContent("this is content");
+
+		//when
+		adminSignUp("s20062", "1234", "jihwan");
+		System.out.println("========= saved =========");
+		adminLogin("s20062", "1234");
+		improvementService.create(improvementDto);
+
+
 	}
 }
